+++ jupyter_tikz/jupyter_tikz.py
@@ -1,6 +1,7 @@
 """Jupyter TikZ is an IPython Cell and Line Magic for rendering TeX/TikZ outputs in Jupyter Notebooks."""
 
 import os
+import re
 import subprocess
 import sys
 import tempfile
@@ -55,6 +56,14 @@ class TexDocument:
         """Returns the full LaTeX code to render."""
         return self._code
 
+    @property
+    def tikz_code(self) -> str | None:
+        pattern = r"\\begin\{tikzpicture\}.*?\\end\{tikzpicture\}"
+        match = re.search(pattern, self.full_latex, re.DOTALL)
+        if match:
+            return match.group(0)
+        return None
+
     @staticmethod
     def _arg_head(arg, limit=60) -> str:
         if type(arg) == str:
@@ -90,24 +99,14 @@ class TexDocument:
         """
         return self._code
 
-    @staticmethod
-    def _modify_texinputs(current_dir: str) -> dict[str, str]:
-        env = os.environ.copy()
-        # TEXINPUTS is a environment variable that tells LaTeX where to look for files
-        # https://tex.stackexchange.com/questions/410350/texinputs-on-windows
-        if "TEXINPUTS" in env:
-            env["TEXINPUTS"] = current_dir + os.pathsep + env["TEXINPUTS"]
-        else:
-            env["TEXINPUTS"] = "." + os.pathsep + current_dir + os.pathsep * 2
-            # note that the trailing double pathsep will insert the standard
-            # search path (otherwise we would lose access to all packages)
-            # TEXINPUTS=.;C:\Users\user\Documents\LaTeX\local\\;
+    def _clearup_latex_garbage(self, keep_temp) -> None:
+        if not (keep_temp):  # F
+            files = Path().glob(f"{self.__hash__()}.*")
+            for file in files:
+                if file.exists():
+                    file.unlink()
 
-        return env
-
-    def _run_command(
-        self, command: str, working_dir: str, full_err: bool = False, **kwargs
-    ) -> int:
+    def _run_command(self, command: str, full_err: bool = False, **kwargs) -> int:
 
         result = subprocess.run(
             command,
@@ -115,7 +114,6 @@ class TexDocument:
             capture_output=True,
             text=True,
             check=False,
-            cwd=working_dir,
             **kwargs,
         )
         if result.returncode != 0:
@@ -125,37 +123,9 @@ class TexDocument:
             print(err_msg, file=sys.stderr)
         return result.returncode
 
-    def save(
-        self,
-        dest: str | None,
-        src: str | None = None,
-        format: Literal["svg", "png", "code"] = "code",
-    ) -> str | None:
-        """Save the TikZ or LaTeX code to a file.
-
-        Args:
-            dest: The destination file path.
-            src: The source file path if the format is a image ("svg" or "png") or code source string if the format is "code".
-            format: The format to save the file. It can be "svg", "png", or "code". If the format is "code", it saves to a text file.
-
-        Raises:
-            ValueError: If the format is not valid.
-            ValueError: If the format is a image ("svg" or "png") and the source file path is not provided.
-
-        Returns:
-            str | None: The destination file absolute path. None if an error occurs.
-        """
-        if dest is None:
-            return None
-
-        if format not in ["svg", "png", "code"]:
-            raise ValueError(
-                f"`{format}` is not a valid format. Valid formats are `svg`, `png`, and `code`."
-            )
-
-        if format != "code" and src is None:
-            raise ValueError("src must be provided when format is not code.")
-
+    def _save(
+        self, dest: str, ext: Literal["tikz", "tex", "png", "svg", "pdf"]
+    ) -> None:
         dest_path = Path(dest)
 
         if os.environ.get("JUPYTER_TIKZ_SAVEDIR"):
@@ -165,24 +135,19 @@ class TexDocument:
 
         dest_path = dest_path.resolve()
 
-        if format == "svg" and (dest_path.suffix != ".svg"):
-            dest_path = dest_path.with_suffix(dest_path.suffix + ".svg")
-        elif format == "png" and (dest_path.suffix != ".png"):
-            dest_path = dest_path.with_suffix(dest_path.suffix + ".png")
-        elif format == "code" and (not dest_path.suffix):
-            dest_path = dest_path.with_suffix(".tex")
+        if dest_path.suffix != f".{ext}":
+            dest_path = dest_path.with_suffix(dest_path.suffix + f".{ext}")
 
         dest_path.parent.mkdir(parents=True, exist_ok=True)
 
-        if format == "code":
-            if src is None:
-                src = self._code
-            dest_path.write_text(src, encoding="utf-8")
+        if ext == "tikz":
+            if not self.tikz_code:
+                raise ValueError("No TikZ code to save.")
+            dest_path.with_suffix(".tikz").write_text(self.tikz_code, encoding="utf-8")
         else:
-            if src is not None:
-                src_path = Path(src).resolve()
-                copy(src_path, dest_path)
-        return f"{dest_path}"
+            Path(str(self.__hash__())).with_suffix(f".{ext}").replace(
+                dest_path.with_suffix(f".{ext}")
+            )
 
     def run_latex(
         self,
@@ -190,8 +155,13 @@ class TexDocument:
         tex_args: str | None = None,
         rasterize: bool = False,
         full_err: bool = False,
+        keep_temp: bool = False,
         save_image: str | None = None,
         dpi: int = 96,
+        grayscale: bool = False,
+        save_tex: str | None = None,
+        save_tikz: str | None = None,
+        save_pdf: str | None = None,
     ) -> Image | SVG | None:
         """Run the LaTeX program to render the LaTeX code.
 
@@ -200,29 +170,30 @@ class TexDocument:
             tex_args: Arguments to pass to the TeX program.
             rasterize: Output a rasterized image (PNG) instead of SVG.
             full_err: Print the full error message when an error occurs. If False, it prints only the last 20 lines.
+            keep_temp: Keep temporary LaTeX files.
             save_image: Save the output image to file.
             dpi: DPI to use when rasterizing the image.
+            grayscale: Set grayscale to a rasterized image.
+            save_tex: Save the full LaTeX code to file.
+            save_tikz: Save the TikZ code to file.
+            save_pdf: Save the output PDF to file.
 
         Returns:
             Image | SVG | None: The rendered image. None if an error occurs.
         """
-        current_dir = os.getcwd()  # get current working directory
+        try:
 
-        with tempfile.TemporaryDirectory() as working_dir:
-            tex_path = Path(working_dir) / "tikz.tex"
-
-            env = self._modify_texinputs(current_dir)
-            output_stem = tex_path.parent.resolve() / tex_path.stem
-
+            tex_path = Path().resolve() / f"{self.__hash__()}.tex"
             tex_path.write_text(self.full_latex, encoding="utf-8")
 
             tex_command = tex_program
             if tex_args:
                 tex_command += f" {tex_args}"
-            tex_command += f" {tex_path.resolve()}"
+            tex_command += f" {tex_path}"
 
-            res = self._run_command(tex_command, working_dir, full_err, env=env)
+            res = self._run_command(tex_command, full_err=full_err)
             if res != 0:
+                self._clearup_latex_garbage(keep_temp)
                 return None
 
             image_format = "svg" if not rasterize else "png"
@@ -234,24 +205,44 @@ class TexDocument:
 
             pdftocairo_command = f"{pdftocairo_path} -{image_format}"
             if rasterize:
-                pdftocairo_command += f" -singlefile -transp -r {dpi}"
+                pdftocairo_command += (
+                    f" -singlefile -{'gray' if grayscale else 'transp'} -r {dpi}"
+                )
 
-            pdftocairo_command += f" {output_stem}.pdf"
+            pdftocairo_command += f" {tex_path.with_suffix('.pdf')}"
             pdftocairo_command += (
-                f" {output_stem}.svg" if not rasterize else f" {output_stem}"
+                f" {tex_path.with_suffix('.svg')}"
+                if not rasterize
+                else f" {tex_path.parent / tex_path.stem}"
             )
-            res = self._run_command(pdftocairo_command, working_dir, full_err)
+            res = self._run_command(pdftocairo_command, full_err=full_err)
+
             if res != 0:
+                self._clearup_latex_garbage(keep_temp)
                 return None
 
-            if not rasterize:
-                if save_image:
-                    self.save(save_image, f"{output_stem}.svg", "svg")
-                return display.SVG(f"{output_stem}.svg")
-            else:
-                if save_image:
-                    self.save(save_image, f"{output_stem}.png", "png")
-                return display.Image(f"{output_stem}.png")
+            image = (
+                display.Image(tex_path.with_suffix(".png"))
+                if rasterize
+                else display.SVG(tex_path.with_suffix(".svg"))
+            )
+
+            if save_image:
+                self._save(save_image, image_format)
+            if save_tex:
+                self._save(save_tex, "tex")
+            if save_pdf:
+                self._save(save_pdf, "pdf")
+            if save_tikz and self.tikz_code:
+                self._save(save_tikz, "tikz")
+
+            self._clearup_latex_garbage(keep_temp)
+
+            return image
+        except Exception as e:
+            raise e
+        finally:
+            self._clearup_latex_garbage(keep_temp)
 
     def _render_jinja(self, ns) -> None:
         try:
@@ -489,11 +480,23 @@ _ARGS = {
         "desc": "DPI to use when rasterizing the image",
         "example": "`--dpi=300`",
     },
+    "gray": {  # New
+        "short-arg": "g",
+        "dest": "gray",
+        "type": bool,
+        "desc": "Set grayscale to a rasterized image",
+    },
     "full-err": {
         "short-arg": "e",
         "dest": "full_err",
         "type": bool,
         "desc": "Print the full error message when an error occurs",
+    },
+    "keep-temp": {  # New
+        "short-arg": "k",
+        "dest": "keep_temp",
+        "type": bool,
+        "desc": "Keep temporary LaTeX files",
     },
     "tex-program": {
         "short-arg": "tp",
@@ -517,13 +520,29 @@ _ARGS = {
         "type": bool,
         "desc": "Do not compile the TeX code",
     },
-    "save-text": {
+    "save-tikz": {  # New
         "short-arg": "s",
+        "dest": "save_tikz",
+        "type": str,
+        "default": None,
+        "desc": "Save the TikZ code to file",
+        "example": "`-s filename.tikz`",
+    },
+    "save-tex": {  # Changed
+        "short-arg": "st",
         "dest": "save_tex",
         "type": str,
         "default": None,
-        "desc": "Save the TikZ or LaTeX code to file",
-        "example": "`-s filename.tikz`",
+        "desc": "Save full LaTeX code to file",
+        "example": "`-st filename.tex`",
+    },
+    "save-pdf": {  # New
+        "short-arg": "sp",
+        "dest": "save_pdf",
+        "type": str,
+        "default": None,
+        "desc": "Save PDF file",
+        "example": "`-sp filename.pdf`",
     },
     "save-image": {
         "short-arg": "S",
@@ -721,16 +740,18 @@ class TikZMagics(Magics):
                 tex_args=self.args.tex_args,
                 rasterize=self.args.rasterize,
                 full_err=self.args.full_err,
+                keep_temp=self.args.keep_temp,
+                save_tikz=self.args.save_tikz,
+                save_tex=self.args.save_tex,
+                save_pdf=self.args.save_pdf,
                 save_image=self.args.save_image,
                 dpi=self.args.dpi,
+                grayscale=self.args.gray,
             )
             if image is None:
                 return None
 
         if self.args.save_var:
             local_ns[self.args.save_var] = str(self.tex_obj)
-
-        if self.args.save_tex:
-            self.saved_path = self.tex_obj.save(self.args.save_tex, format="code")
 
         return image